Refactored mention to follow other prosemirror node
