+++ components/common/CharmEditor/components/mention/components/index.ts
@@ -1,0 +1,7 @@
+import MentionNode from './MentionNode';
+import MentionSuggest from './MentionSuggest';
+
+export {
+  MentionNode,
+  MentionSuggest
+};
