+++ components/common/CharmEditor/components/mention/index.ts
@@ -1,0 +1,6 @@
+export * from './components';
+export * from './mention.constants';
+export * from './mention.plugins';
+export * from './mention.interfaces';
+export * from './mention.specs';
+export * from './mention.utils';
