+++ components/common/CharmEditor/components/mention/mention.constants.ts
@@ -1,0 +1,4 @@
+export const mentionSuggestMarkName = 'mentionSuggest';
+export const mentionTrigger = '@';
+export const mentionPluginKeyName = 'mentionSuggest';
+export const mentionNodeName = 'mention';
