+++ components/common/CharmEditor/components/mention/mention.interfaces.ts
@@ -1,0 +1,4 @@
+import { SuggestPluginState } from 'lib/prosemirror/interfaces';
+
+export interface MentionPluginState extends SuggestPluginState {
+}
